Use PHPMailer 6
Changed all occurrences of $this->LE to static::$LE except in amazonSesSend() which is changed to remove trailing white space, as phpmailer itself does.

Tidied-up the visibility of class properties, those which are additional as opposed to inherited, and removed some that appear to be unused.

Removed some obsolete code that supported earlier versions of phpmailer
add_attachment()
add_html_image()