Adding the ability to set the max processing time for the processqueue page
The current processqueue page, even when called from a command line
(using curl for example), will stop processing the queue after 60
seconds. This change lets you specify the max processing time in the GET
request so that the processqueue page can be used outside of a web
browser for batch processing.