Better support for non-latin strings in select dropdown of search subscribers page
I am using phpList along with attributes in the Greek language and in the search subscribers section the filtering dropdown which shows the available attributes doesn't correctly truncate Greek strings due to the substr command.

I propose to change it with the function shortenTextDisplay (defined in lib.php) which uses the corresponding mb_ versions for substr etc. if available.